+++ src/constants/colors-config.js
@@ -43,7 +43,9 @@ var _default = {
   // Red
   error2: '#ffe5ec',
   // Light Red
-  error3: '#ffb2c7' // Semi Light Red
-
+  error3: '#ffb2c7',
+  // Semi Light Red
+  warning: '#8a6d3b',
+  warning2: '#fcf8e3'
 };
 exports.default = _default;